requirements.txt
python-dotenv
streamlit 
langchain-community
langchain-groq
langchain-openai
langchain-ollama
langchain-gemini
